+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 public interface ItemService {
 
-    ItemDto addItem(ItemDto itemDto, int userId);
+    ItemDto addItem(ItemDto itemDto, long userId);
 
-    ItemDto getItemById(int id);
+    ItemDto getItemById(long id);
 
-    ItemDto updateItem(ItemDto itemDto, int userId);
+    ItemDto updateItem(ItemDto itemDto, long userId);
 
-    List<ItemDto> getItemsByOwnerId(int id);
+    List<ItemDto> getItemsByOwnerId(long id);
 
     List<ItemDto> findItemByText(String text);
 }
